+++ src/transforms.h
@@ -8,12 +8,12 @@
 
 struct transform final {
     vec3 offset;
-    double sin_theta;
-    double cos_theta;
+    float sin_theta;
+    float cos_theta;
 
     constexpr transform() = default;
 
-    transform(double angleDegrees, vec3 offset) noexcept;
+    transform(float angleDegrees, vec3 offset) noexcept;
 
     point3 applyForward(point3 p) const noexcept;
     aabb applyForward(aabb) const noexcept;
